Implement Contact Us API
In this commit, a new function handleSendMessage has been introduced to manage the process of sending messages via a POST request. The function is designed to be triggered by an event, in this case, the submission of a form. Upon execution, it prevents the default form submission behavior using the e.preventDefault() call.

Within the function, an attempt is made to send the message using the Axios library. The API endpoint URL is currently set to a placeholder value ("http://127.0.0.1:8000/api/contact-us"), and it is recommended to update this URL to the actual endpoint.

Upon receiving a response from the API, the code checks the HTTP status. If the status is 200 (indicating a successful request), the response data is logged to the console, and a success alert is displayed using the Swal library. This alert provides feedback to the user about the successful message submission, along with a timer indicating how long the alert will be displayed. Additionally, the email and message fields are cleared to reset the form.

If the response status is not 200, an error alert is displayed, indicating that the message sending process has failed. The error message from the response data is logged, and an error alert similar to the success alert is shown, with the same timer behavior.

In case an error occurs during the API call (handled within the catch block), a general error alert is displayed. This alert informs the user about the occurrence of an error, regardless of whether it was related to the API call or not. The error message is logged to the console, and the error alert is displayed with the timer as previously described.

It is important to note that while the current implementation provides functional error handling and user feedback, the placeholder API endpoint should be replaced with the actual API endpoint URL before deploying this code to a production environment.