Create slider and fix bug
In this commit, sliders have been created and bugs fixed.

Previously there was a bug where the size of the slider was too stretched which did not match the design in figma. In addition, there is also a bug in the button for next or back to the previous slide where the position does not fit in the center when on mobile display. Both of these bugs have been resolved.